# speedyGo-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8020",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/gestionStation/rental/updateRental": {
      "put": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "updateRental",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/bookingSeats/{id}": {
      "get": {
        "tags": [
          "booking-seats-controller"
        ],
        "operationId": "getBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSeats"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "booking-seats-controller"
        ],
        "operationId": "updateBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSeats"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "booking-seats-controller"
        ],
        "operationId": "deleteBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/gestionStation/rental/affecterVehicleRentalToRental": {
      "post": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "affecterVehicleRentalToRental",
        "parameters": [
          {
            "name": "idVehicleRental",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idRental",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/gestionStation/rental/affecterStationStartRental": {
      "post": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "affecterStationStartRental",
        "parameters": [
          {
            "name": "rentalId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idStartStation",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/gestionStation/rental/affecterStationDestinationRental": {
      "post": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "affecterStationDestinationRental",
        "parameters": [
          {
            "name": "rentalId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idDestination",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/gestionStation/rental/affecterRentalToCustomer": {
      "post": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "affecterRentalToCustomer",
        "parameters": [
          {
            "name": "emailCustomer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idRental",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/gestionStation/rental/addRental": {
      "post": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "addRental",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/bookingSeats/create": {
      "post": {
        "tags": [
          "booking-seats-controller"
        ],
        "operationId": "createBooking",
        "parameters": [
          {
            "name": "carpoolingId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "seatsRequested",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSeats"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestionStation/rental/getAllRental": {
      "get": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "getAllRental",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestionStation/rental/deleteRental": {
      "delete": {
        "tags": [
          "rental-controller"
        ],
        "operationId": "deleteRental",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Carpooling": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "departLocation": {
            "type": "string"
          },
          "arrivalLocation": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "seatsAvailable": {
            "type": "integer",
            "format": "int32"
          },
          "pricePerSeat": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
...
      "LoyaltyProgram": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "partnerLoyaltyProgram": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "chat": {
            "$ref": "#/components/schemas/Chat"
          }
        }
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Parcel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "typeParcel": {
            "type": "string",
            "enum": [
              "ClOTHING",
              "FRAGILE",
              "FOOD"
            ]
          },
          "customerParcel": {
            "$ref": "#/components/schemas/User"
          },
          "driverParcel": {
            "$ref": "#/components/schemas/User"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
...
      "Rental": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "cost": {
            "type": "number"
          },
          "distanceTraveled": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ReservationTrip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "taxi": {
            "$ref": "#/components/schemas/Taxi"
          },
          "carpooling": {
            "$ref": "#/components/schemas/Carpooling"
          },
          "trip": {
            "$ref": "#/components/schemas/Trip"
          },
          "payment": {
            "$ref": "#/components/schemas/Payment"
          },
          "carpoolingUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "taxiUser": {
            "$ref": "#/components/schemas/User"
          },
          "driver": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "customerReview": {
            "$ref": "#/components/schemas/User"
          },
          "driverReview": {
            "$ref": "#/components/schemas/User"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
...
      "SupportTicket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "customerSupportTicket": {
            "$ref": "#/components/schemas/User"
          },
          "driverSupportTicket": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          },
          "driver": {
            "$ref": "#/components/schemas/User"
          },
          "pickupLocation": {
            "type": "string"
          },
          "dropoffLocation": {
            "type": "string"
          },
          "distance": {
            "type": "number",
            "format": "float"
          },
          "fare": {
            "type": "number",
            "format": "float"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "DRIVER",
              "PARTNER",
              "CUSTOMER"
            ]
          },
          "partnerType": {
            "type": "string",
            "enum": [
              "RESTAURANT",
              "STORE",
              "PHARMACY"
            ]
          },
          "address": {
            "type": "string"
          },
          "licence": {
            "type": "string"
          },
          "insurance": {
            "type": "string"
          },
          "criminalRecord": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "availability": {
            "type": "boolean"
          },
          "reservationTripsCarpooling": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationTrip"
            }
          },
          "reservationTripsTaxi": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationTrip"
            }
          },
          "reservationTripDriver": {
            "$ref": "#/components/schemas/ReservationTrip"
          },
          "tripsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trip"
            }
          },
          "tripsDriver": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trip"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "reviewsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "reviewsDriver": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "supportTicketsDriver": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportTicket"
            }
          },
          "supportTicketsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportTicket"
            }
          },
          "loyaltyProgramsPartner": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoyaltyProgram"
            }
          },
          "loyaltyProgramsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoyaltyProgram"
            }
          },
          "parcelsDriver": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Parcel"
            }
          },
          "parcelsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Parcel"
            }
          },
          "eventPromotionsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event_Promotion"
            }
          },
          "eventPromotionsPartner": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event_Promotion"
            }
          },
          "chatsSender": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chat"
            }
          },
          "chatsReceived": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chat"
            }
          },
          "productsCustomer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "productsPartner": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "stationManagment": {
            "$ref": "#/components/schemas/Station"
          },
          "rentals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          },
          "postListReceiver": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "postSender": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "BookingSeats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "seatsRequested": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "carpooling": {
            "$ref": "#/components/schemas/Carpooling"
          }
        }
      },
...
    },
    "securitySchemes": {
      "keycloak": {
        "type": "oauth2",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:9090/realms/speedyGo/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:9090/realms/speedyGo/protocol/openid-connect/token",
            "scopes": {

            }
          }
        }
      }
    }
  }
}